pipeline/pyproject.toml
[build-system]
requires = ["flit_core==3.9.0"]
build-backend = "flit_core.buildapi"

[project]
name = "boilercv_pipeline"
version = "0.0.0"
description = "Data pipeline."
dependencies = [
    "astroid>=3.3.1",
    "boilercine>=2024.1.1",
    "boilercore>=2024.1.3",
    "cappa>=0.22.2",
    "cmasher>=1.7.2",
    "colorcet>=3.0.1",
    "copykitten>=1.1.1",
    "cyclopts>=2.6.1",
    "imageio[pyav]>=2.34.1",
    "loguru>=0.7.0",
    "matplotlib>=3.7.2",
    "numpy>=1.24.4",
    "pandas[hdf5,performance]>=2.2.2",
    "pillow>=10.3.0",
    "ploomber-engine>=0.0.30",
    "pyarrow>=14.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "pyqtgraph>=0.13.3",
    "ruamel.yaml>=0.17.32",
    "scikit-image>=0.23.1",
    "scipy>=1.11.1",
    "seaborn>=0.13.2",
    "sparklines>=0.4.2",
    "stopit>=1.1.2",
    "sympy>=1.12",
    "tomlkit>=0.12.4",
    "tqdm>=4.66.3",
    "watchfiles>=0.22.0",
    "xarray[accel,io,parallel]>=2023.7.0",
    "pyyaml>=6.0.2",
]
[project.scripts]
"boilercv_pipeline" = "boilercv_pipeline.__main__:main"
[project.optional-dependencies]
# * Minimal CV flavor needed, but allow `boilercv` to coexist with different flavors
cv = ["opencv-python-headless>=4.10.0.84"]
[tool.fawltydeps]
deps = ["pipeline/pyproject.toml"]
code = ["pipeline/boilercv_pipeline"]
ignore_undeclared = ["boilercv", "boilercv_docs", "PySide6"]
ignore_unused = [
    "pyarrow", # Used in `boilercv_pipeline.__init__`
]
[tool.fawltydeps.custom_mapping]
boilercine = ["boilercine"]
boilercore = ["boilercore"]
